requirements.txt
streamlit
numpy
tensorflow
pydicom
Pillow
